+++ tests/tests_cycle/test_cases.py
@@ -52,6 +52,11 @@ TEST_INVERSE = [
     (Cycle(1, 3, 4, 2), Cycle(2, 4, 3, 1)),
     (Cycle(2, 3, 1, 5, 4), Cycle(4, 5, 1, 3, 2)),
 ]
+TEST_INVERSIONS = [
+    (Cycle(1, 2, 3), []),
+    (Cycle(1, 3, 4, 2), [(2, 4), (3, 4)]),
+    (Cycle(1, 2, 5, 4, 3), [(3, 4), (3, 5), (4, 5)]),
+]
 TEST_IS_CONJUGATE_ERROR = [(Cycle(1, 2, 3), Cycle(1, 2, 3), NotImplementedError, "Method")]
 TEST_IS_DERANGEMENT = [
     (Cycle(1), False),